// frontend/src/core/api/model/updateMedicalRecordDto.ts
/**
 * Generated by orval v8.0.0 🍺
 * Do not edit manually.
 * Dentizy API
 * API Documentation untuk Sistem Manajemen Klinik Gigi
 * OpenAPI spec version: 1.0
 */

export interface UpdateMedicalRecordDto {
  /**
   * Update bagian Subjektif dari SOAP (keluhan pasien)
   * @maxLength 5000
   * @nullable
   */
  subjektif?: string | null;
  /**
   * Update bagian Objektif dari SOAP (hasil pemeriksaan fisik/lab)
   * @maxLength 5000
   * @nullable
   */
  objektif?: string | null;
  /**
   * Update bagian Assessment dari SOAP (diagnosis sementara/lanjutan)
   * @maxLength 5000
   * @nullable
   */
  assessment?: string | null;
  /**
   * Update bagian Plan dari SOAP (rencana terapi & tindak lanjut)
   * @maxLength 5000
   * @nullable
   */
  plan?: string | null;
}